BoardService 게시글 상세 조회 매개변수 타입 수정

+++ springPractice/src/main/java/org/example/springpractice/Board/service/BoardService.java
@@ -19,7 +19,7 @@ public class BoardService {
     }
 
     //게시글 상세 조회
-    public BoardDto.BoardRes read(Long id) {
+    public BoardDto.BoardRes read(Integer id) {
         return boardRepository.findById(id).map(BoardDto.BoardRes::from).orElse(null);
     }
